Refatorei funções em classes e terminei o endpoint de signup
