--- modulo5/projeto-iwfs-coockenu/RespostasTeoricas.md
@@ -1,8 +1,0 @@
-/*
-    Exercício 1
-    a) o round ou cost, é um custo númerico, quanto maior mais seguro será a senha e vai demorar mais para rodar, o recomendado e mais usado é 12.
-        o salt é uma string aleatória adicionada ao código para evitar os ataques chamados rainbow table
-
-    Exercício 2
-    a) O de cadastro, para cadastrar novos usuários com a senha já com o hash
-*/
