--- modulo5/projeto-iwfs-coockenu/exercicios_criptografia_userrole.sql
@@ -1,3 +1,0 @@
-ALTER TABLE UserAulaAutenticao ADD COLUMN role VARCHAR(255) DEFAULT "normal" ;
-
-SELECT * FROM UserAulaAutenticao;
