+++ modulo5/projeto-iwfs-coockenu/src/error/EmptyName.ts
@@ -1,0 +1,7 @@
+import { BaseError } from "./BaseError";
+
+export class EmptyName extends BaseError {
+    constructor() {
+        super('Name field cannot be empty', 422)
+    }
+}
