+++ modulo5/projeto-iwfs-coockenu/src/index.ts
@@ -3,6 +3,6 @@ import { User } from "./endpoints/User"
 
 const user: User = new User()
 
-app.get("/usuario/login", user.login)
-app.get("/usuario/profile", user.pegarDadosUsuario)
-app.post("/usuario/cadastrar", user.createUser)
+app.get("/login", user.login)
+app.get("/user/profile", user.pegarDadosUsuario)
+app.post("/signup", user.createUser)
